+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.0.4 - In progress
 ### Changed
 - Reorder LCMS fields.
+- `atacseq` to `scatacseq`.
 
 ## [v0.0.3](https://github.com/hubmapconsortium/ingest-validation-tools/tree/v0.0.3) - 2020-05-04
 ### Added